// src/common/enums/checkin.enum.ts
export enum CheckinType {
  ENTRY = 'entry',
  EXIT = 'exit',
}

export enum CheckinStatus {
  NORMAL = 'normal',
  LATE = 'late',
  EARLY_LEAVE = 'early_leave',
  OVERTIME = 'overtime',
  ABNORMAL = 'abnormal',
}

export enum PersonType {
  STAFF = 'staff',
  SUPPLIER = 'supplier',
  VISITOR = 'visitor',
  CONTRACTOR = 'contractor',
}